requirements/requirements-dev.txt
addict==2.2.0
alabaster==0.7.12
amqp==2.3.2
appdirs==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
Babel==2.6.0
billiard==3.5.0.4
black==18.9b0
celery==4.2.1
certifi==2018.10.15
cffi==1.12.2
channels==2.1.7
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
daphne==2.2.5
Django==2.1.3
django-extensions==2.1.4
django-filter==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.9.0
djoser==1.3.1
docutils==0.14
hyperlink==18.0.0
idna==2.7
imagesize==1.1.0
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
Markdown==3.0.1
MarkupSafe==1.1.0
numpy==1.15.4
packaging==18.0
pandas==0.23.4
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
Pygments==2.3.0
PyHamcrest==1.9.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
redis==3.0.1
requests==2.20.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
toml==0.10.0
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
xgboost==0.80
zope.interface==4.6.0
